+++ README.md
@@ -47,7 +47,7 @@ _Get a local copy up and running by following these simple steps._
 4. Type in your terminal the command `ruby ISBN.rb` to run the program
 
 ### Run Unit Tests(Rspec)
-1. Run the command `rspec`or `rspec ./spec/ISBN_spec.rb`
+1. Run the command `rspec` or `rspec ./spec/ISBN_spec.rb`
 
         OR
 1. Go into the folder containing the project by typing `cd Hyperiondev-THT`
